backend/pyproject.toml
[project]
name = "juanfer-seguros-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[tool.uv.workspace]
members = ["."]
